big_data_shenyang/requirements.txt
pyquery==1.4.1
PyMySQL==0.10.1
opencv_python==4.4.0.44
requests==2.24.0
selenium==3.141.0
xmltodict==0.12.0
pytesseract==0.3.6
pandas==0.19.2
lxml==4.5.2
Pillow==7.2.0
